+++ examples/user.go
@@ -2,7 +2,7 @@ package examples
 
 //go:generate ../bin/gobuilder -type=User -file=user.go -output=user_builder.go
 type User struct {
-	ID    int
-	Name  string
+	ID    int    `required:"true" default:"1"`
+	Name string  `omitempty:"true"`
 	Email string
 }